adding text moving footer back into main project
