+++ lib/app/widgets/widgets.dart
@@ -1,0 +1,1 @@
+export 'footer.dart';
