+++ lib/l10n/l10n.dart
@@ -1,4 +1,3 @@
-// coverage:ignore-file
 import 'package:flutter/widgets.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
