+++ packages/blog_ui/lib/src/theme/blog_text_styles.dart
@@ -25,4 +25,10 @@ abstract class BlogTextStyles {
     fontSize: 32,
     fontWeight: FontWeight.w800,
   );
+
+  /// Style for the footer text.
+  static const TextStyle footerTextStyle = TextStyle(
+    fontSize: 14,
+    fontWeight: FontWeight.w400,
+  );
 }
