+++ packages/blog_ui/lib/src/widgets/widgets.dart
@@ -2,4 +2,3 @@ export 'author_avatar.dart';
 export 'author_tile.dart';
 export 'blog_card.dart';
 export 'blog_card_hero.dart';
-export 'footer.dart';
