+++ packages/blog_ui/pubspec.yaml
@@ -17,8 +17,3 @@ dev_dependencies:
   mocktail: ^1.0.0
   network_image_mock: ^2.0.1
   very_good_analysis: ^5.1.0
-
-flutter:
-  assets:
-    - assets/
-    - assets/images/
