+++ pubspec.yaml
@@ -36,3 +36,6 @@ dev_dependencies:
 flutter:
   uses-material-design: true
   generate: true
+  assets:
+    - assets/
+    - assets/images/
